+++ src/vsix/UI/Converters/BoolToPinConverter.cs
@@ -21,7 +21,7 @@
 
         public override object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (!(value is bool boolValue))
+            if (value is not bool boolValue)
                 return value;
 
             if (Reverse)