+++ src/vsix/UI/ViewModels/NewsItems/NewsItemViewModel.cs
@@ -1,14 +1,23 @@
 ﻿using System;
 
-using CommunityToolkit.Mvvm.Messaging;
 using CommunityToolkit.Mvvm.Input;
+using CommunityToolkit.Mvvm.Messaging;
 
 namespace StartPagePlus.UI.ViewModels.NewsItems
 {
     using Messages;
 
+    using Options.Pages;
+
     public partial class NewsItemViewModel : ViewModelBase
     {
+        //private readonly bool _showTooltip;
+
+        //public NewsItemViewModel()
+        //{
+
+        //}
+
         public string Title { get; set; }
 
         public string Description { get; set; }
@@ -16,6 +25,9 @@ namespace StartPagePlus.UI.ViewModels.NewsItems
         public string Link { get; set; }
 
         public DateTime Date { get; set; }
+
+        public bool ShowTooltip
+            => NewsItemsOptions.Instance.ShowNewsItemTooltip;
 
         //---